+++ src/test/java/com/tgerstel/service/TransferServiceTest.java
@@ -2,6 +2,7 @@ package com.tgerstel.service;
 
 import static org.junit.jupiter.api.Assertions.*;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.Optional;
@@ -48,12 +49,12 @@ class TransferServiceTest {
 		dateTime = LocalDate.now();		
 		userActual = new User("Bob", "sobob@a.com", "hardpass", 13);
 		userActual.setId(1L);
-		receipt = new Receipt(dateTime, 1200.0f, 200.0f, null, null, "Customer", "Me", "for example", userActual);
+		receipt = new Receipt(dateTime, BigDecimal.valueOf(1200), BigDecimal.valueOf(200.0), null, null, "Customer", "Me", "for example", userActual);
 		receipt.setId(22L);
 		user2 = new User("Rob", "roby@am.com", "hardpass", 12);
-		receiptWithUser2 = new Receipt(dateTime, 1000.0f, 200.0f, null, null, "Sansumg", "Me", "for example", user2);
-		transfer = new Transfer(TransferType.IN_TRANSFER, 1200.0f, "Customer", "Me", dateTime,	null, receipt, userActual);
-		transferWithReceipt2 = new Transfer(TransferType.IN_TRANSFER, 1200.0f, "Customer", "Me", dateTime,	null, receiptWithUser2, userActual);
+		receiptWithUser2 = new Receipt(dateTime, BigDecimal.valueOf(1000), BigDecimal.valueOf(200.0), null, null, "Sansumg", "Me", "for example", user2);
+		transfer = new Transfer(TransferType.IN_TRANSFER,BigDecimal.valueOf(1200), "Customer", "Me", dateTime,	null, receipt, userActual);
+		transferWithReceipt2 = new Transfer(TransferType.IN_TRANSFER, BigDecimal.valueOf(1200), "Customer", "Me", dateTime,	null, receiptWithUser2, userActual);
 	}
 	
 
@@ -72,13 +73,13 @@ class TransferServiceTest {
 		
 		assertAll(
 				() -> assertEquals(dateTime, transferReturned.getDate()),
-				() -> assertEquals(1200f, transferReturned.getAmount()),
+				() -> assertEquals(BigDecimal.valueOf(1200), transferReturned.getAmount()),
 				() -> assertEquals("Bob", transferReturned.getUser().getUsername()),
 				() -> assertNotNull(transferReturned.getUser()),
 				() -> assertEquals(TransferType.IN_TRANSFER, transferReturned.getTransferType()),				
 				() -> assertEquals("Customer", transferReturned.getFrom()),
 				() -> assertEquals("Me", transferReturned.getTo()),
-				() -> assertEquals(1200f, transferReturned.getReceipt().getAmount()),
+				() -> assertEquals(BigDecimal.valueOf(1200), transferReturned.getReceipt().getAmount()),
 				() -> assertNotNull(transferReturned.getReceipt()),
 				() -> assertEquals("Customer", transferReturned.getReceipt().getClient())	
 		);
@@ -129,7 +130,7 @@ class TransferServiceTest {
 		
 		assertAll(			
 				() -> assertEquals(dateTime, transferReturned.getDate()),
-				() -> assertEquals(1200f, transferReturned.getAmount()),
+				() -> assertEquals(BigDecimal.valueOf(1200), transferReturned.getAmount()),
 				() -> assertEquals("Bob", transferReturned.getUser().getUsername()),
 				() -> assertEquals(TransferType.IN_TRANSFER, transferReturned.getTransferType())
